Added a lot (nothing works atm)
I'd never push something that doesn't build & run, but the actual progress isn't really displayed in game yet.

Also learned the hard way that any cpp/hpp files we add need to manually be added to the cmakelists or it won't compile.

There's a lot of comments, have fun with that.