Tile system start (#29)
* added a speed cap

This doesn't work towards our goal of a runescape clone at all :)

* Split MapEntity class

The MapEntity class was split into MapObjects & MapEntities, the idea being that having players, npcs, walls, items all sharing the same class wasn't good for the future.
MapObjects covers everything that is rendered on the map (& currently the map itself)
MapEntites extends MapObjects & has variables and functions for movement.

The point of this & the next step is to have velocity, speed, and movement tracked by the player object instead of it being calculated in main.

For future reference;
If it doesn't move, it's a MapObject (walls, ground items, background).
If it moves, or can be moved, it's a MapEntity.

FURTHMORE I over comment everything 1. So I can remember and 2. Incase anyone else jumps in on the project. I recommend dumbing down all of your code through comments so the less fortunate (me) doesn't have to stare at your code for 2 hours just to understand it.

* Added a lot (nothing works atm)

I'd never push something that doesn't build & run, but the actual progress isn't really displayed in game yet.

Also learned the hard way that any cpp/hpp files we add need to manually be added to the cmakelists or it won't compile.

There's a lot of comments, have fun with that.

* 5_15_25 push

commiting what I apparently left from before

* important comments

sea plus plus