Valid function
Mostly code cleanup. A function was made to check if a game object is valid, currently only checks if a map and an ID exist. Felt easier to have a function check for validity rather than having to manually flag things as valid.
It would be best though to have a function later that validates all gameobjects on init or something.

Also, note how despite never being "added" to the map, the player is still rendered, can move, and interact with map objects.