Display & map update (#30)
* added a speed cap

This doesn't work towards our goal of a runescape clone at all :)

* Split MapEntity class

The MapEntity class was split into MapObjects & MapEntities, the idea being that having players, npcs, walls, items all sharing the same class wasn't good for the future.
MapObjects covers everything that is rendered on the map (& currently the map itself)
MapEntites extends MapObjects & has variables and functions for movement.

The point of this & the next step is to have velocity, speed, and movement tracked by the player object instead of it being calculated in main.

For future reference;
If it doesn't move, it's a MapObject (walls, ground items, background).
If it moves, or can be moved, it's a MapEntity.

FURTHMORE I over comment everything 1. So I can remember and 2. Incase anyone else jumps in on the project. I recommend dumbing down all of your code through comments so the less fortunate (me) doesn't have to stare at your code for 2 hours just to understand it.

* Added a lot (nothing works atm)

I'd never push something that doesn't build & run, but the actual progress isn't really displayed in game yet.

Also learned the hard way that any cpp/hpp files we add need to manually be added to the cmakelists or it won't compile.

There's a lot of comments, have fun with that.

* 5_15_25 push

commiting what I apparently left from before

* important comments

sea plus plus

* display changes (wip)

* display change

I somehow rewrote everything & it still works.
In hindsight very little of the display class was actually changed, I kind of just worked myself in a circle & stated adding things back that I removed when things didn't work.

* Valid function

Mostly code cleanup. A function was made to check if a game object is valid, currently only checks if a map and an ID exist. Felt easier to have a function check for validity rather than having to manually flag things as valid.
It would be best though to have a function later that validates all gameobjects on init or something.

Also, note how despite never being "added" to the map, the player is still rendered, can move, and interact with map objects.

* big todo (WIP)

I break things while writing about how I broke them & can't fix it.

* map vector test (WIP)

Functions for a vector in Map created, Display updated to use that vector.

* I broke git

donald trump please save me

* I fixed git

"Baseline" in vcpkg.config changed when I updated VS. I found the new one by following a tutorial to setup vcpkg/cmake in a new project. No guarantee what I have now will work for everyone or just me.

* Restored collision

Collision detection is now working properly. Also removed all unused/broken code for tiles.

* 6/20 changes

3rd times the charm.
"GetObject" for the allObjects list was removed entirely. this was easier than trying to rewrite it to work with a std:list.
Immobile flag for MapObjects was removed. It should be self evident if something that shouldn't move is moving.

* update gitignore